// WebApplication5/Areas/Admin/AdminApp/src/app/core/models/trips/trip.model.ts
export class Trip {
    id: string;
    name: string;
    content: string;
    url: string;
    imagePath: string;
    startDate: Date;
    endDate: Date;
    price: number;
    priceText: string;
    sale: number;
    hasSale: boolean;
    days: number;
    nights: number;
    groupSize: number;
    tourDates: string;
    map: string;
    seoDescription: string;
    seoKeywords: string;
    destinationId: string;
    activityId: string;
    typeOfTripId: string;
}